If an sld body is available in a geofile, try to publish it into the remote server
Add two createStyle() variants:
- one variant using the default workspace and using an empty sld body
- one variant taking the workspace, layer and an sld body string
- reuse the latter in the previous createLayer() unique function
All three variants now return a boolean whether the last status code
was 200 or not.

GeoServer returns a 200 code even if the published sld body is invalid,
so GET it after PUT'ing it to ensure it was validated. Fallback to
the default style otherwise, as it was done before.